Remove json-key from innermost functions
json-key is the hash table key used to look up the value of the
response from Google.  But putting it in the innermost async layer
embeds knowledge of the response that it doesn't need.  Removing it
saves a parameter in the API as well.  The right place for the key is
in the callback functions that are specific to the type of data being
fetched, such as elg-call-label-list-endpoint-callback.